DP climate input and MPI_read

+++ shared.f90
@@ -24,7 +24,7 @@ REAL, DIMENSION (ntimes,nland) :: source ! 0.4 GB
 REAL, DIMENSION (nland) :: source_lat
 REAL, DIMENSION (nland) :: source_larea ! km2
 REAL, DIMENSION (nlon,nlat,ntimes) :: clm_in ! 1.5 GB
-REAL, ALLOCATABLE, DIMENSION (:,:) :: result ! 0.1 GB
+REAL (DP), ALLOCATABLE, DIMENSION (:,:) :: result ! 0.1 GB
 REAL, ALLOCATABLE, DIMENSION (:) :: lon_chunk
 REAL, ALLOCATABLE, DIMENSION (:) :: lat_chunk
 REAL, ALLOCATABLE, DIMENSION (:,:,:) :: tmp ! K ! 0.59 (30 yr over 20 processes)
